Updates after first successful Travis build
This is the first working build from TravisCI. It does not include
dependencies (they must already be installed on the home system) but it
does have a working version of waveletEncoder and a few CGAL test cases.

README has been updated to reflect the changes. Branch will not be
deleted--use in future for testing dependency building.