Moved file conversion and added extraction script
DICOM data from DCMCCT is coming as a filesystem tree of .zip files.
Clearly, we don't want to manually extract all the data. The
build_data_tree.sh script automates the process of extracting the .zip
files (with various formats), conversion to RAWIV, and metadata dumping.

The last two parts still need to be implemented.