Update dicom2rawiv to produce axial scans only
Updated convert_all so that it will only output metadata/rawiv for
series where the change in patient Z (not image Z) is larger than change
in patient X or patient Y (left and back, respectively in the RCS coord
system).

This should cause only the CT scans that are axial in nature to be
stored.