Added precurate binary
You'll need to turn on the LIBIGL flag for this to work (can use my
software or download your own--notes in the CMakeLists.txt file). Works
even with side-plates. Usage is:

  precurate <input.inr> <output.inr> <isovalue> [debug]

I've used iso value of 600 and it works nicely. The debug option is
anything, and will use libigl to display the resulting mesh. Running
time is fairly quick (<5 minutes) on a full volume input.