Add working global pipeline scripts
Now have an option to build a set of scripts that drive the global image
pipeline. This also provides guidance on how to run these programs.

Fixed a bug in volimg_convert that would cause the script to incorrectly
determine the output location when specifying an outformat but not an
outname.